requirements.txt
python-telegram-bot==22.1
pyserial==3.5